api: ssize_t read/write and reject oversized uid/gid
Two narrowing fixes:

- filsys_read/filsys_write returned int while the backends return ssize_t
  and the facade cast the result back; make them ssize_t and drop the casts.

- A create/chown with a modern uid that doesn't fit the 16-bit on-disk
  field silently wrapped (70000 -> 4464).  Reject it with -EINVAL in
  create/mkdir/mknod/symlink and per-field in chown, the way mknod already
  rejects major/minor > 255.

+++ filsys.h
@@ -171,10 +171,10 @@ int filsys_readdir(filsys_t *fs, const char *path, filsys_dirent_t **ents,
 
 /* ---- file data ----------------------------------------------------------- */
 
-int filsys_read(filsys_t *fs, const char *path, void *buf, size_t size,
-                off_t off);
-int filsys_write(filsys_t *fs, const char *path, const void *buf, size_t size,
-                 off_t off);
+ssize_t filsys_read(filsys_t *fs, const char *path, void *buf, size_t size,
+                    off_t off);
+ssize_t filsys_write(filsys_t *fs, const char *path, const void *buf, size_t size,
+                     off_t off);
 /* Read a symlink's target (no trailing NUL) into buf; returns the byte count.
  * -ENOSYS if the edition predates symlinks, -EINVAL if path is not a symlink. */
 ssize_t filsys_readlink(filsys_t *fs, const char *path, char *buf, size_t size);
